% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_estimation.R
\name{alpha_estimation}
\alias{alpha_estimation}
\title{alpha estimation procedure}
\usage{
alpha_estimation(X, K = 3, Rs = 1:15, opts = list())
}
\arguments{
\item{X}{a quantitative data set}

\item{K}{K-fold cross validation}

\item{Rs}{the searching range of the number of components}

\item{opts}{a list contains the setting for the algorithm. \itemize{
\item tol_obj: tolerance for relative change of hist_obj, default:1E-6;
\item maxit: max number of iterations, default: 1000;
}}
}
\value{
This function returns a list contains \itemize{
\item alphas_mean: the mean of the K estimated alphas
\item alphas_std: the std of the K estimated alphas
\item R_CV: the number of PCs with minimum cross validation error
\item cvErrors: K-fold cross validation errors
}
}
\description{
This function will estimate the noise level (alpha is the variation parameter)
of a quantitative data set using the PCA model. The rank of the PCA model
is selected using a missing value based cross validation procedure. The
details of this function can be found in \url{https://arxiv.org/abs/1902.06241}.
}
\examples{
\dontrun{alpha_estimation(X,K=3,Rs = 1:15,opts=list())}

}
