% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{svd_mis}
\alias{svd_mis}
\title{A svd algorithm with the option for missing values.}
\usage{
svd_mis(X, R, opts)
}
\arguments{
\item{X}{a quantitative data set}

\item{R}{the number of PCs}

\item{opts}{a list contains the setting for the algorithm. \itemize{
\item tol_obj: tolerance for relative change of hist_obj, default:1E-6;
\item maxit: max number of iterations, default: 1000;
}}
}
\value{
This function returns a list contains \itemize{
\item U: the left singular vectors
\item S: a column vector contains the singular values
\item V: the right singular vectors
\item iter: the number of iterations used
\item diagnose: records hist_obj and rel_obj
}
}
\description{
This function implemented a MM algorithm to fit a svd on a quantitaive data
set with missing values. The details of this function can be found
in \url{https://arxiv.org/abs/1902.06241}.
}
\examples{
\dontrun{svd_mis(X,R=3,opts=list())}

}
